Move frss elements to separate folder
